Revert recent "git merge <msg> HEAD <commit>..." deprecation
This reverts commit 1026457ac2995b403ea7fec94c22a8baa673d5f9 "git-pull.sh:
Fix call to git-merge for new command format" and

commit 4f42c035d01f76823f93b711483e348000e7b1b0 "git-merge: a deprecation
notice of the ancient command line syntax".

They caused a "git pull" (without any arguments, and without any local
commits---only to update to the other side) to warn that commit log
message is ignored because the merge resulted in a fast-forward.

Another possible solution is to add an extra option to "git merge" so that
"git pull" can tell it that the message given is not coming from the end
user (the canned message is passed just in case the merge resulted in a
non-ff and caused commit), but I think it is easier _not_ to deprecate the
old syntax.

Signed-off-by: Junio C Hamano <gitster@pobox.com>